+++ src/vallm/__init__.py
@@ -13,4 +13,4 @@ __all__ = [
     "VallmSettings",
 ]
 
-__version__ = "0.1.3"
+__version__ = "0.1.4"